Utilized ControllersTestsBase in all 4 controller test classes. Moved EventsController tests to different file.
